pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = ""
authors = [
    {name = "田中結那（たなかゆいな）",email = "yuina.tnk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
